repr: catch spike up to current main (re-migrate drifted refs + tests)
Rebased the int-unification spike onto current main. The base spike commit
cherry-picked cleanly; this commit re-applies the mechanical follow-on work
against the new base and migrates the references that landed in main after the
original branch point (#37155, #37177, #36776 count-aware-aggregates, #36984):

- production: signed-sum fold + SumInt16/32 narrow (relation/func.rs), the
  generate_series-unoptimized literal (sql/func.rs), and the generate_series
  collapse (projection_pushdown.rs: ReprScalarType::Int64 -> Int, and the
  literal_i64 closure's int2/4/8 arms collapse to one Datum::Int arm).
- test/bench/doctest: ~292 Datum::IntN -> Datum::Int/UInt sites, plus the
  PropDatum::Int32 narrowings and an i64-typed literal in casts.rs.

Workspace builds clean under --all-targets.

+++ src/arrow-util/src/reader.rs
@@ -1040,8 +1040,8 @@ mod tests {
 
         packer.extend([
             Datum::True,
-            Datum::Int32(42),
-            Datum::UInt64(10000),
+            Datum::Int(42),
+            Datum::UInt(10000),
             Datum::Float32(OrderedFloat::from(-1.1f32)),
             Datum::String("hello world"),
             Datum::Bytes(b"1010101"),
@@ -1052,7 +1052,7 @@ mod tests {
                 serde_json::json!({"code": 200, "email": "space_monkey@materialize.com"}),
             )
             .expect("failed to pack JSON");
-        packer.push_list([Datum::UInt32(200), Datum::UInt32(300)]);
+        packer.push_list([Datum::UInt(200), Datum::UInt(300)]);
 
         let null_row = Row::pack(vec![Datum::Null; 9]);
 
@@ -1248,11 +1248,11 @@ mod tests {
             .push_range(Range::new(Some((
                 RangeLowerBound {
                     inclusive: true,
-                    bound: Some(Datum::Int32(1)),
+                    bound: Some(Datum::Int(1)),
                 },
                 RangeUpperBound {
                     inclusive: true,
-                    bound: Some(Datum::Int32(10)),
+                    bound: Some(Datum::Int(10)),
                 },
             ))))
             .unwrap();
@@ -1266,11 +1266,11 @@ mod tests {
             .push_range(Range::new(Some((
                 RangeLowerBound {
                     inclusive: false,
-                    bound: Some(Datum::Int32(5)),
+                    bound: Some(Datum::Int(5)),
                 },
                 RangeUpperBound {
                     inclusive: false,
-                    bound: Some(Datum::Int32(15)),
+                    bound: Some(Datum::Int(15)),
                 },
             ))))
             .unwrap();
