+++ src/expr/src/interpret.rs
@@ -1719,7 +1719,7 @@ mod tests {
                         func: Eq.into(),
                         expr1: Box::new(MirScalarExpr::column(0)),
                         expr2: Box::new(MirScalarExpr::literal_ok(
-                            Datum::Int32(1727694505),
+                            Datum::Int(1727694505),
                             ReprScalarType::Int,
                         )),
                     },
@@ -1732,7 +1732,7 @@ mod tests {
         let relation = ReprRelationType::new(vec![ReprScalarType::Int.nullable(true)]);
         let arena = RowArena::new();
         let mut interpreter = ColumnSpecs::new(&relation, &arena);
-        interpreter.push_column(0, ResultSpec::value(Datum::Int32(-1294725158)));
+        interpreter.push_column(0, ResultSpec::value(Datum::Int(-1294725158)));
         let spec = interpreter.mfp_filter(&mfp);
         assert!(spec.range.may_fail());
     }
@@ -1758,7 +1758,7 @@ mod tests {
     #[mz_ore::test]
     fn test_eval_range() {
         // Example inspired by the tumbling windows temporal filter in the docs
-        let period_ms = MirScalarExpr::literal_ok(Datum::Int64(10), ReprScalarType::Int);
+        let period_ms = MirScalarExpr::literal_ok(Datum::Int(10), ReprScalarType::Int);
         let expr = MirScalarExpr::CallBinary {
             func: Gte.into(),
             expr1: Box::new(MirScalarExpr::CallUnmaterializable(
