+++ src/transform/src/case_literal.rs
@@ -368,14 +368,14 @@ mod tests {
 
     /// Helper: build an i64 literal.
     fn lit_i64(v: i64) -> MirScalarExpr {
-        MirScalarExpr::literal_ok(Datum::Int64(v), ReprScalarType::Int)
+        MirScalarExpr::literal_ok(Datum::Int(v), ReprScalarType::Int)
     }
 
     /// Wrap a scalar expression in a `Map` over a constant to allow applying the transform.
     fn wrap_in_map(scalar: MirScalarExpr) -> MirRelationExpr {
         MirRelationExpr::Map {
             input: Box::new(MirRelationExpr::constant(
-                vec![vec![Datum::Int64(0)]],
+                vec![vec![Datum::Int(0)]],
                 ReprRelationType::new(vec![ReprColumnType {
                     scalar_type: ReprScalarType::Int,
                     nullable: false,
@@ -486,20 +486,20 @@ mod tests {
         // Spot-check evaluation.
         let arena = mz_repr::RowArena::new();
         assert_eq!(
-            result.eval(&[Datum::Int64(0)], &arena).unwrap(),
-            Datum::Int64(0)
+            result.eval(&[Datum::Int(0)], &arena).unwrap(),
+            Datum::Int(0)
         );
         assert_eq!(
-            result.eval(&[Datum::Int64(32)], &arena).unwrap(),
-            Datum::Int64(3200)
+            result.eval(&[Datum::Int(32)], &arena).unwrap(),
+            Datum::Int(3200)
         );
         assert_eq!(
-            result.eval(&[Datum::Int64(63)], &arena).unwrap(),
-            Datum::Int64(6300)
+            result.eval(&[Datum::Int(63)], &arena).unwrap(),
+            Datum::Int(6300)
         );
         assert_eq!(
-            result.eval(&[Datum::Int64(999)], &arena).unwrap(),
-            Datum::Int64(-1)
+            result.eval(&[Datum::Int(999)], &arena).unwrap(),
+            Datum::Int(-1)
         );
     }
 
@@ -513,20 +513,20 @@ mod tests {
         let arena = mz_repr::RowArena::new();
 
         // Input = 1 → output should be 10
-        let out = result.eval(&[Datum::Int64(1)], &arena).unwrap();
-        assert_eq!(out, Datum::Int64(10));
+        let out = result.eval(&[Datum::Int(1)], &arena).unwrap();
+        assert_eq!(out, Datum::Int(10));
 
         // Input = 2 → output should be 20
-        let out = result.eval(&[Datum::Int64(2)], &arena).unwrap();
-        assert_eq!(out, Datum::Int64(20));
+        let out = result.eval(&[Datum::Int(2)], &arena).unwrap();
+        assert_eq!(out, Datum::Int(20));
 
         // Input = 99 → output should be 0 (els)
-        let out = result.eval(&[Datum::Int64(99)], &arena).unwrap();
-        assert_eq!(out, Datum::Int64(0));
+        let out = result.eval(&[Datum::Int(99)], &arena).unwrap();
+        assert_eq!(out, Datum::Int(0));
 
         // Input = NULL → output should be 0 (els, since NULL = x is falsy)
         let out = result.eval(&[Datum::Null], &arena).unwrap();
-        assert_eq!(out, Datum::Int64(0));
+        assert_eq!(out, Datum::Int(0));
     }
 
     #[mz_ore::test]
@@ -575,15 +575,15 @@ mod tests {
         let arena = mz_repr::RowArena::new();
         assert_eq!(
             result.eval(&[Datum::String("a")], &arena).unwrap(),
-            Datum::Int64(1)
+            Datum::Int(1)
         );
         assert_eq!(
             result.eval(&[Datum::String("b")], &arena).unwrap(),
-            Datum::Int64(2)
+            Datum::Int(2)
         );
         assert_eq!(
             result.eval(&[Datum::String("z")], &arena).unwrap(),
-            Datum::Int64(0)
+            Datum::Int(0)
         );
     }
 }
